Remove TimeGuard in favor of built-in socket timeout configuration.

+++ Java/Example/src/Client.java
@@ -11,7 +11,7 @@ public class Client {
 
         String host = "127.0.0.1";
         int port = 8877;
-        long timeout = 10000; // 10 seconds
+        int timeout = 10000; // 10 seconds
 
         ProcBridge pb = new ProcBridge(host, port, timeout);
